requirements.txt
asgiref==3.8.1
Django==5.1.5
sqlparse==0.5.3
djangorestframework==3.15.2
python-dotenv==1.0.1
openai==1.59.8